refactor(google_sheets): migrate to new MCP server architecture
Migrate the google_sheets MCP server to follow the new modular
architecture pattern established by hubspot and other servers.

- Create metadata.ts with createToolsRecord for type-safe tool definitions
- Create helpers.ts with withAuth patterns for Google API clients
- Create tools/index.ts with ToolHandlers for exhaustive handler mapping
- Create index.ts with registerTool pattern for server creation
- Update constants.ts to use metadata field instead of inline serverInfo
- Update servers/index.ts to import from new location
